requirements.txt
typing
numpy>=1.24.4
pandas==2.0.3
polars-lts-cpu>=0.20.19
seaborn==0.13.0
kaggle>=1.6.12
matplotlib==3.8.2
scikit-learn==1.3.2
plotly==5.20.0
arviz>=0.15.1
causalpy==0.2.2
mapie==0.8.3
pyarrow==16.0.0
pystan==3.8.0